+++ assets/js/admin.js
@@ -7,3 +7,4 @@ import './publish-drafts';
 import './tool-ajax-form';
 import './not-synced';
 import './system-reports';
+import './sync-outdated';