+++ includes/init/Install.php
@@ -33,24 +33,11 @@ class Install
 	 */
     public function run_install()
     {
-        self::check_wc_activation();
+
         add_action('plugins_loaded', array(__CLASS__, 'awca_update_plugin_version'));
 
         $this->set_options();
     }
-
-	/**
-	 * Check if WooCommerce is activated.
-	 *
-	 * @since    1.0.0
-	 */
-	private static function check_wc_activation()
-	{
-		if (!class_exists('WooCommerce')) {
-			deactivate_plugins(plugin_basename(__FILE__));
-			wp_die('پلاگین فوق نیاز به فعال بودن ووکامرس دارد لطفا ابتدا ووکامرس را نصب و فعال کنید. <br><a href="' . admin_url('plugins.php') . '">برگشت به صفحه افزونه</a>');
-		}
-	}
 
 
 	/**